requirements.txt
altair==4.1.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.15.6
botocore==1.18.6
Brlapi==0.6.6
cachetools==4.1.1
certifi==2018.1.18
cffi==1.14.3
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
httplib2==0.9.2
idna==2.6
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.3
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.2
oauth==1.0.1
olefile==0.45.1
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==1.0.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydeck==0.5.0b1
Pygments==2.7.1
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.3.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.15.0
streamlit==0.67.1
system-service==0.3
systemd-python==234
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
validators==0.18.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
wadllib==1.3.2
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.2.0
zope.interface==4.3.2
